% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_lpms.r
\name{loss_lpms}
\alias{loss_lpms}
\title{Causal binary loss function model for Bayesian linear probability models}
\usage{
loss_lpms(
  posterior,
  pi,
  md = 0,
  mu = 0,
  unit = 1,
  cp = 1,
  ce = c(rep(101:10000)/100)
)
}
\arguments{
\item{posterior}{A numeric vector of posterior samples for the estimated effect of the intervention}

\item{pi}{A numeric value or vector for the baseline likelihood of the undesirable outcome}

\item{md}{A numeric value for the minimum desired effect size on the probability scale; default = 0}

\item{mu}{A numeric value for the minimum undesired effect size on the probability scale; default = 0}

\item{unit}{A numeric value for the size of the unit change of interest in the intervention variable; default = 1}

\item{cp}{A numeric value for the cost of the intervention; default = 1}

\item{ce}{A numeric value or vector for the cost(s) of the undesirable outcome; default = c(rep(101:10000)/100)}
}
\value{
A list having two elements: the first element is a data frame that has the expected losses under the intervention ("losses_i"), those under no intervention ("losses_noi"), and the ratio of the cost of the intervention to the cost of the undesirable outcome ("cost_ratio"); the second element is the ggplot object that plots the first element
}
\description{
This function implements the causal binary loss function model (Suzuki 2022)
for Bayesian linear probability models.
}
\details{
This function implements the causal binary loss function model (Suzuki 2022) for Bayesian
linear probability models. It computes expected losses under an intervention (in causal inference
senses) and those under no intervention (i.e., the status quo), across different ratios of the
cost of the intervention to the cost of an undesirable outcome.

For simplicity, the function computes these ratios by taking a fixed value for the cost of an
intervention, while taking a vector of various values for the cost of an undesirable outcome.

It is possible to specify the minimum desirable and undesirable effect sizes on the probability
scale (\code{md} and \code{mu} respectively). If \code{md}=0, the strict inequality md<0 is assumed;
otherwise the inequality <= and >= is assumed.

One caveat in the setup of linear probability models is that the binary dependent variable
must be coded such that the value of 1 means an undesirable outcome and the value of 0 means
a desirable outcome

If a greater value of the intervention variable captures an intervention meant to increase
the likelihood of an undesirable outcome, you must specify a negative value for \code{unit},
the size of the unit change of interest, in the function. In this way, a decrease in the
intervetion variable captures an intervention meant to reduce the likelihood of an undesirable
outcome

If the variable of an intervention is a continuous / ordered variable, you can specify
\code{unit}, the size of the unit change of interest, other than 1 (which is the default value).

If the average treatment effect is the quantity of interest and there are more than one independent
variable used in a linear probability model, the vector should be created that is the sum of all
covariates plus the posterior mean of their coefficients, and this should be used as an input
for \code{pi}. The loss function will compute the predicted probability of the outcome for each
observation and then will use their mean (the missing values in the vector, if any, will be
removed).

For a full example from estimating posterior samples to using the package, please see
\href{https://akisatosuzuki.github.io/bayesdtm.html}{https://akisatosuzuki.github.io/bayesdtm.html}.

For a theoretical rationale for using Bayesian decision-theoretic models such as the causal binary
loss function model, please see the paper:

\href{https://doi.org/10.1080/2330443X.2022.2050328}{Suzuki, Akisato. 2022. "Policy Implications of
Statistical Estimates: A General Bayesian Decision-Theoretic Model for Binary Outcomes."
Statistics and Public Policy, https://doi.org/10.1080/2330443X.2022.2050328.}

If you use this package, please cite the following items:

Suzuki, Akisato. 2022. "Policy Implications of Statistical Estimates: A General Bayesian
Decision-Theoretic Model for Binary Outcomes." Statistics and Public Policy,
https://doi.org/10.1080/2330443X.2022.2050328.

Suzuki, Akisato. 2022. "bayesdtm: Bayesian Decision-Theoretic Models to Compute Expected Losses
under an Intervention and the Status Quo." R package version 1.0.0.
}
\section{Author(s)}{

Author & Maintainer: Akisato Suzuki (\email{akisato.suzuki@gmail.com})
}

\examples{
\dontrun{
# Compute expected losses
loss_lpm(posterior = posterior, pi = pi)

# Compute expected losses with specifying the minimum desired effect size
# as a probability difference of 0.1
loss_lpm(posterior = posterior, pi = pi, md = 0.1)

# If a value of one in the binary variable captures an intervention meant
# to increase the likelihood of an undesirable outcome, specify -1 in unit
loss_lpm(posterior = posterior, pi = pi, unit = -1)

# Change the ratio of the costs
loss_lpm(posterior = posterior, pi = pi, cp = 2, ce = rep(5:10))
}
}
