% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_logit.r
\name{loss_logit}
\alias{loss_logit}
\title{Compute expected losses based on Bayesian logistic regression}
\usage{
loss_logit(posterior, pi, md = 0, mu = 0)
}
\arguments{
\item{posterior}{A vector of posterior samples for the estimated effect of an intervention}

\item{pi}{A baseline log odds value of an undesirable event}

\item{md}{A minimum desired effect size in a log odds scale; default = 0}

\item{mu}{A minimum undesired effect size in a log odds scale; default = 0}
}
\value{
A list having two elements: the first element is a data frame that has the expected losses given an intervention ("losses_i"), those given no intervention ("losses_noi"), and the ratio of the cost of the intervention to the cost of the undesirable event ("cost_ratio"); the second element is a ggplot object that plots the first element
}
\description{
This function implements a Bayesian decision-theoretic model for outputs from Bayesian
logistic regression.
}
\details{
This function implements a Bayesian decision-theoretic model for outputs from Bayesian
logistic regression. It computes expected losses given an intervention (in causal inference
senses) and those without it, across different ratios of the cost of the intervention to
the cost of an undesirable event. More specifically, in the loss function the cost of an
intervention is fixed at 1 while the cost of an undesirable event is changed from 1 to 100.
So the ratio of the costs is from 0.01 to 1, and there are more data points the smaller
the ratio.

There are two caveats in the setup of logistic regression. The binary dependent variable
must be coded so that the value of 1 means an undesirable event and the value of 0 means
a desirable event. Similarly, the variable of an intervention must be coded so that a
larger value means an effect to reduce the likelihood of an undesirable event.

For a full example from estimating posterior samples to using the package, please see
\href{https://akisatosuzuki.github.io/bayesdtm.html}{https://akisatosuzuki.github.io/bayesdtm.html}.

For a theoretical rationale for using this model, please see the following paper:

\href{https://arxiv.org/abs/2008.10903}{Suzuki, Akisato. 2020. "Policy Implications of
Statistical Estimates: A General Bayesian Decision-Theoretic Model for Binary Outcomes."
arXiv:2008.10903 [stat.ME]. https://arxiv.org/abs/2008.10903.}
}
\section{Author(s)}{

Author & Maintainer: Akisato Suzuki (\email{akisato.suzuki@gmail.com})
}

\examples{
\dontrun{
# Compute expected losses
loss_logit(posterior = posterior, pi = pi)

# Compute expected losses with specifying the minimum desired effect size
# as an odds ratio of 0.8
loss_logit(posterior = posterior, pi = pi, md = log(0.8))
}
}
